Get slider working in IE
IE does not register an ‘input’ event until the user lets go of the
slider, whereas it does trigger a ‘change’ event on every change. If I
remember correctly, this is actually the spec, and no one else follows
it. In any case, you gotta do it both ways for now.

cc’ing @michaelbjames, the kind of fact you’d rather not know!