Use Network.Websockets.forkPingThread.
This seems to fix a condition where the client would no longer receive
updates. The notifier was working, but the data never showed up on the
client side of the websocket. I suspect this is due to the keepalive
thread, but I couldn't nail it down precisely. Also, this is my first
ever PR in Haskell, so it's very likely that something about this is
wrong.