work around import bug
which caused errors of the form:

Error in frontend/SideBar.elm:

Error on line 34, column 14 to 38:
Could not find variable 'Controls.permitSwapMailbox.signal'.

Looks like the prefix 'Controls.permitSwapMailbox' is not in scope. Is it spelled correctly?
Is it imported correctly?

You can read about how imports work at the following address:
<http://elm-lang.org/learn/Syntax.elm#modules>