Bundle all static files with Template Haskell, clean up index code
This change starts using Template Haskell (TH) to bundle the necessary
compiled Elm code in the binary. This will mean we no longer need an
ELM_HOME and can avoid https://github.com/kevva/elm-bin/issues/22
entirely.

This also introduces the cabal-build.py build script which touches the
TH file if any of the Elm resources have been modified since lasts we
compiled.

It also cleans up the new index code a decent amount. We took out all
the icons in favor of font-awesome which can again be bundled with the
binary instead of living in ELM_HOME.